EVG-14861: refactor packages to flatten interface definitions
