+++ ecs/pod_creator.go
@@ -1,0 +1,37 @@
+package ecs
+
+import (
+	"context"
+
+	"github.com/evergreen-ci/cocoa"
+	"github.com/pkg/errors"
+)
+
+// BasicECSPodCreator provides an ECSPodCreator implementation to create
+// ECS pods.
+type BasicECSPodCreator struct {
+	client cocoa.ECSClient
+	vault  cocoa.Vault
+}
+
+// NewBasicECSPodCreator creates a helper to create pods backed by ECS.
+func NewBasicECSPodCreator(c cocoa.ECSClient, v cocoa.Vault) (*BasicECSPodCreator, error) {
+	if c == nil {
+		return nil, errors.New("missing client")
+	}
+	return &BasicECSPodCreator{
+		client: c,
+		vault:  v,
+	}, nil
+}
+
+// CreatePod creates a new pod backed by ECS.
+func (m *BasicECSPodCreator) CreatePod(ctx context.Context, opts ...*cocoa.ECSPodCreationOptions) (cocoa.ECSPod, error) {
+	return nil, errors.New("TODO: implement")
+}
+
+// CreatePodFromExistingDefinition creates a new pod backed by ECS from an
+// existing definition.
+func (m *BasicECSPodCreator) CreatePodFromExistingDefinition(ctx context.Context, def cocoa.ECSTaskDefinition, opts ...*cocoa.ECSPodExecutionOptions) (cocoa.ECSPod, error) {
+	return nil, errors.New("TODO: implement")
+}
