+++ ecs/pod_creator_test.go
@@ -1,0 +1,12 @@
+package ecs
+
+import (
+	"testing"
+
+	"github.com/evergreen-ci/cocoa"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestECSPodCreatorInterface(t *testing.T) {
+	assert.Implements(t, (*cocoa.ECSPodCreator)(nil), &BasicECSPodCreator{})
+}
