+++ ecs_pod_test.go
@@ -1,174 +1,12 @@
 package cocoa
 
 import (
-	"context"
-	"os"
 	"testing"
-	"time"
 
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/evergreen-ci/cocoa/awsutil"
-	"github.com/evergreen-ci/cocoa/secret"
 	"github.com/evergreen-ci/utility"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestECSPod(t *testing.T) {
-	assert.Implements(t, (*ECSPod)(nil), &BasicECSPod{})
-
-	checkAWSEnvVarsForECSAndSecretsManager(t)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	for tName, tCase := range map[string]func(ctx context.Context, t *testing.T, c ECSClient){
-		"InfoIsPopulated": func(ctx context.Context, t *testing.T, c ECSClient) {
-			res := NewECSPodResources().SetTaskID("task_id")
-			stat := Starting
-			opts := NewBasicECSPodOptions().SetClient(c).SetResources(*res).SetStatus(stat)
-
-			p, err := NewBasicECSPod(opts)
-			require.NoError(t, err)
-
-			info, err := p.Info(ctx)
-			require.NoError(t, err)
-			assert.Equal(t, *res, info.Resources)
-			assert.Equal(t, stat, info.Status)
-		},
-	} {
-		t.Run(tName, func(t *testing.T) {
-			tctx, tcancel := context.WithTimeout(ctx, 30*time.Second)
-			defer tcancel()
-
-			hc := utility.GetHTTPClient()
-			defer utility.PutHTTPClient(hc)
-			awsOpts := awsutil.NewClientOptions().
-				SetHTTPClient(hc).
-				SetCredentials(credentials.NewEnvCredentials()).
-				SetRole(os.Getenv("AWS_ROLE")).
-				SetRegion(os.Getenv("AWS_REGION"))
-
-			c, err := NewBasicECSClient(*awsOpts)
-			require.NoError(t, err)
-
-			tCase(tctx, t, c)
-		})
-	}
-}
-
-func TestBasicECSPodOptions(t *testing.T) {
-	t.Run("NewBasicECSPodOptions", func(t *testing.T) {
-		opts := NewBasicECSPodOptions()
-		require.NotZero(t, opts)
-		assert.Zero(t, *opts)
-	})
-	t.Run("SetClient", func(t *testing.T) {
-		c, err := NewBasicECSClient(*awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1"))
-		require.NoError(t, err)
-		opts := NewBasicECSPodOptions().SetClient(c)
-		assert.Equal(t, c, opts.Client)
-	})
-	t.Run("SetVault", func(t *testing.T) {
-		c, err := secret.NewBasicSecretsManagerClient(*awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1"))
-		require.NoError(t, err)
-		v := secret.NewBasicSecretsManager(c)
-		opts := NewBasicECSPodOptions().SetVault(v)
-		assert.Equal(t, v, opts.Vault)
-	})
-	t.Run("SetResources", func(t *testing.T) {
-		res := NewECSPodResources().SetTaskID("id")
-		opts := NewBasicECSPodOptions().SetResources(*res)
-		require.NotZero(t, opts.Resources)
-		assert.Equal(t, *res, *opts.Resources)
-	})
-	t.Run("SetStatus", func(t *testing.T) {
-		stat := Starting
-		opts := NewBasicECSPodOptions().SetStatus(stat)
-		require.NotNil(t, opts.Status)
-		assert.Equal(t, stat, *opts.Status)
-	})
-	t.Run("Validate", func(t *testing.T) {
-		t.Run("EmptyIsInvalid", func(t *testing.T) {
-			opts := NewBasicECSPodOptions()
-			assert.Error(t, opts.Validate())
-		})
-		t.Run("AllFieldsPopulatedIsValid", func(t *testing.T) {
-			awsOpts := awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1")
-			ecsClient, err := NewBasicECSClient(*awsOpts)
-			require.NoError(t, err)
-			smClient, err := secret.NewBasicSecretsManagerClient(*awsOpts)
-			require.NoError(t, err)
-			v := secret.NewBasicSecretsManager(smClient)
-			res := NewECSPodResources().SetTaskID("id")
-			opts := NewBasicECSPodOptions().
-				SetClient(ecsClient).
-				SetVault(v).
-				SetResources(*res).
-				SetStatus(Starting)
-			assert.NoError(t, opts.Validate())
-		})
-		t.Run("MissingClientIsInvalid", func(t *testing.T) {
-			awsOpts := awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1")
-			smClient, err := secret.NewBasicSecretsManagerClient(*awsOpts)
-			require.NoError(t, err)
-			v := secret.NewBasicSecretsManager(smClient)
-			res := NewECSPodResources().SetTaskID("id")
-			opts := NewBasicECSPodOptions().
-				SetVault(v).
-				SetResources(*res).
-				SetStatus(Starting)
-			assert.Error(t, opts.Validate())
-		})
-		t.Run("MissingVaultIsValid", func(t *testing.T) {
-			awsOpts := awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1")
-			ecsClient, err := NewBasicECSClient(*awsOpts)
-			require.NoError(t, err)
-			res := NewECSPodResources().SetTaskID("id")
-			opts := NewBasicECSPodOptions().
-				SetClient(ecsClient).
-				SetResources(*res).
-				SetStatus(Starting)
-			assert.NoError(t, opts.Validate())
-		})
-		t.Run("MissingResourcesIsInvalid", func(t *testing.T) {
-			awsOpts := awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1")
-			smClient, err := secret.NewBasicSecretsManagerClient(*awsOpts)
-			require.NoError(t, err)
-			v := secret.NewBasicSecretsManager(smClient)
-			res := NewECSPodResources()
-			opts := NewBasicECSPodOptions().
-				SetVault(v).
-				SetResources(*res).
-				SetStatus(Starting)
-			assert.Error(t, opts.Validate())
-		})
-		t.Run("BadResourcesIsInvalid", func(t *testing.T) {
-			awsOpts := awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1")
-			smClient, err := secret.NewBasicSecretsManagerClient(*awsOpts)
-			require.NoError(t, err)
-			v := secret.NewBasicSecretsManager(smClient)
-			opts := NewBasicECSPodOptions().
-				SetVault(v).
-				SetStatus(Starting)
-			assert.Error(t, opts.Validate())
-		})
-		t.Run("MissingStatusIsInvalid", func(t *testing.T) {
-			awsOpts := awsutil.NewClientOptions().SetCredentials(credentials.NewEnvCredentials()).SetRegion("us-east-1")
-			ecsClient, err := NewBasicECSClient(*awsOpts)
-			require.NoError(t, err)
-			smClient, err := secret.NewBasicSecretsManagerClient(*awsOpts)
-			require.NoError(t, err)
-			v := secret.NewBasicSecretsManager(smClient)
-			res := NewECSPodResources().SetTaskID("id")
-			opts := NewBasicECSPodOptions().
-				SetClient(ecsClient).
-				SetVault(v).
-				SetResources(*res)
-			assert.Error(t, opts.Validate())
-		})
-	})
-}
 
 func TestPodSecret(t *testing.T) {
 	t.Run("NewPodSecret", func(t *testing.T) {
