+++ internal/awsutil/log.go
