+++ internal/testcase/ecs_client.go
@@ -1,0 +1,191 @@
+package testcase
+
+import (
+	"context"
+	"os"
+	"testing"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/ecs"
+	"github.com/evergreen-ci/cocoa"
+	"github.com/evergreen-ci/cocoa/internal/testutil"
+	"github.com/evergreen-ci/utility"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+// ECSClientTestCase represents a test case for a cocoa.ECSClient.
+type ECSClientTestCase func(ctx context.Context, t *testing.T, c cocoa.ECSClient)
+
+// ECSClientTaskDefinitionTests returns common test cases that a cocoa.ECSClient
+// should support.
+func ECSClientTaskDefinitionTests() map[string]ECSClientTestCase {
+	return map[string]ECSClientTestCase{
+		"RegisterTaskDefinitionFailsWithInvalidInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.RegisterTaskDefinition(ctx, &ecs.RegisterTaskDefinitionInput{})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"DeregisterTaskDefinitionFailsWithInvalidInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.DeregisterTaskDefinition(ctx, &ecs.DeregisterTaskDefinitionInput{})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"RegisterAndDeregisterTaskDefinitionSucceeds": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			registerOut, err := c.RegisterTaskDefinition(ctx, &ecs.RegisterTaskDefinitionInput{
+				ContainerDefinitions: []*ecs.ContainerDefinition{
+					{
+						Command: []*string{aws.String("echo"), aws.String("hello")},
+						Image:   aws.String("busybox"),
+						Name:    aws.String("hello_world"),
+					},
+				},
+				Cpu:    aws.String("128"),
+				Memory: aws.String("4"),
+				Family: aws.String(testutil.NewTaskDefinitionFamily(t.Name())),
+			})
+			require.NoError(t, err)
+			require.NotNil(t, registerOut)
+			require.NotNil(t, registerOut.TaskDefinition)
+			require.NotNil(t, registerOut.TaskDefinition.TaskDefinitionArn)
+			require.NotZero(t, registerOut.TaskDefinition.Status)
+			assert.Equal(t, "ACTIVE", *registerOut.TaskDefinition.Status)
+			require.NotZero(t, registerOut.TaskDefinition.RegisteredAt)
+			assert.NotZero(t, *registerOut.TaskDefinition.RegisteredAt)
+
+			out, err := c.DeregisterTaskDefinition(ctx, &ecs.DeregisterTaskDefinitionInput{
+				TaskDefinition: registerOut.TaskDefinition.TaskDefinitionArn,
+			})
+			require.NoError(t, err)
+			require.NotZero(t, out)
+		},
+	}
+}
+
+// ECSClientRegisteredTaskDefinitionTests returns common test cases that a
+// cocoa.ECSClient should support that rely on a registered task definition.
+func ECSClientRegisteredTaskDefinitionTests(registerIn ecs.RegisterTaskDefinitionInput, registerOut ecs.RegisterTaskDefinitionOutput) map[string]ECSClientTestCase {
+	return map[string]ECSClientTestCase{
+		"RunTaskFailsWithInvalidInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.RunTask(ctx, &ecs.RunTaskInput{})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"DescribeTasksFailsWithInvalidInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.DescribeTasks(ctx, &ecs.DescribeTasksInput{})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"StopTaskFailsWithInvalidInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.StopTask(ctx, &ecs.StopTaskInput{})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"RunTaskFailsWithValidButNonexistentInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.RunTask(ctx, &ecs.RunTaskInput{
+				Cluster:        aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				TaskDefinition: aws.String(utility.RandomString()),
+			})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"DescribeTasksReturnsFailureWithValidButNonexistentInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.DescribeTasks(ctx, &ecs.DescribeTasksInput{
+				Cluster: aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				Tasks:   []*string{aws.String(utility.RandomString())},
+			})
+			assert.NoError(t, err)
+			assert.NotZero(t, out)
+			assert.NotZero(t, out.Failures)
+			assert.Empty(t, out.Tasks)
+		},
+		"StopTaskFailsWithValidButNonexistentInput": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			out, err := c.StopTask(ctx, &ecs.StopTaskInput{
+				Cluster: aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				Task:    aws.String(utility.RandomString()),
+			})
+			assert.Error(t, err)
+			assert.Zero(t, out)
+		},
+		"RegisterSucceedsWithDuplicateTaskDefinition": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			outDuplicate, err := c.RegisterTaskDefinition(ctx, &registerIn)
+			require.NoError(t, err)
+			require.NotZero(t, outDuplicate)
+			require.NotZero(t, outDuplicate.TaskDefinition)
+
+			defer cleanupTaskDefinition(ctx, t, c, outDuplicate)
+
+			assert.True(t, *outDuplicate.TaskDefinition.Revision > *registerOut.TaskDefinition.Revision)
+		},
+		"RunAndStopTaskSucceedsWithRegisteredTaskDefinition": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			require.NotZero(t, registerOut.TaskDefinition.Status)
+			assert.Equal(t, "ACTIVE", *registerOut.TaskDefinition.Status)
+
+			runOut, err := c.RunTask(ctx, &ecs.RunTaskInput{
+				Cluster:        aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				TaskDefinition: registerOut.TaskDefinition.TaskDefinitionArn,
+			})
+			require.NoError(t, err)
+			require.NotZero(t, runOut)
+			require.Empty(t, runOut.Failures)
+			require.NotEmpty(t, runOut.Tasks)
+			assert.Equal(t, runOut.Tasks[0].TaskDefinitionArn, registerOut.TaskDefinition.TaskDefinitionArn)
+
+			out, err := c.StopTask(ctx, &ecs.StopTaskInput{
+				Cluster: aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				Task:    aws.String(*runOut.Tasks[0].TaskArn),
+			})
+			require.NoError(t, err)
+			require.NotZero(t, out)
+			require.NotZero(t, out.Task)
+			assert.Equal(t, runOut.Tasks[0].TaskArn, out.Task.TaskArn)
+		},
+		"DescribeTaskSucceedsWithRunningTask": func(ctx context.Context, t *testing.T, c cocoa.ECSClient) {
+			require.NotZero(t, registerOut.TaskDefinition.Status)
+			assert.Equal(t, "ACTIVE", *registerOut.TaskDefinition.Status)
+
+			runOut, err := c.RunTask(ctx, &ecs.RunTaskInput{
+				Cluster:        aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				TaskDefinition: registerOut.TaskDefinition.TaskDefinitionArn,
+			})
+			require.NoError(t, err)
+			require.NotZero(t, runOut)
+			require.NotEmpty(t, runOut.Tasks)
+
+			defer cleanupTask(ctx, t, c, runOut)
+
+			out, err := c.DescribeTasks(ctx, &ecs.DescribeTasksInput{
+				Cluster: aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+				Tasks:   []*string{aws.String(*runOut.Tasks[0].TaskArn)},
+			})
+			require.NoError(t, err)
+			require.NotZero(t, out)
+			require.NotEmpty(t, out.Tasks)
+			assert.Equal(t, *out.Tasks[0].TaskDefinitionArn, *registerOut.TaskDefinition.TaskDefinitionArn)
+			assert.Equal(t, out.Tasks[0].TaskArn, runOut.Tasks[0].TaskArn)
+		},
+	}
+}
+
+// cleanupTaskDefinition cleans up an existing task definition.
+func cleanupTaskDefinition(ctx context.Context, t *testing.T, c cocoa.ECSClient, out *ecs.RegisterTaskDefinitionOutput) {
+	if out != nil && out.TaskDefinition != nil && out.TaskDefinition.TaskDefinitionArn != nil {
+		out, err := c.DeregisterTaskDefinition(ctx, &ecs.DeregisterTaskDefinitionInput{
+			TaskDefinition: out.TaskDefinition.TaskDefinitionArn,
+		})
+		require.NoError(t, err)
+		require.NotZero(t, out)
+	}
+}
+
+// cleanupTask cleans up a running task.
+func cleanupTask(ctx context.Context, t *testing.T, c cocoa.ECSClient, runOut *ecs.RunTaskOutput) {
+	if runOut != nil && len(runOut.Tasks) > 0 && runOut.Tasks[0].TaskArn != nil {
+		out, err := c.StopTask(ctx, &ecs.StopTaskInput{
+			Cluster: aws.String(os.Getenv("AWS_ECS_CLUSTER")),
+			Task:    aws.String(*runOut.Tasks[0].TaskArn),
+		})
+		require.NoError(t, err)
+		require.NotZero(t, out)
+	}
+}
