+++ mock/mock_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/evergreen-ci/cocoa"
-	"github.com/evergreen-ci/cocoa/secret"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -13,6 +12,6 @@ func TestInterfaces(t *testing.T) {
 	assert.Implements(t, (*cocoa.ECSPod)(nil), &ECSPod{})
 	assert.Implements(t, (*cocoa.ECSClient)(nil), &ECSClient{})
 
-	assert.Implements(t, (*secret.Vault)(nil), &Vault{})
-	assert.Implements(t, (*secret.SecretsManagerClient)(nil), &SecretsManagerClient{})
+	assert.Implements(t, (*cocoa.Vault)(nil), &Vault{})
+	assert.Implements(t, (*cocoa.SecretsManagerClient)(nil), &SecretsManagerClient{})
 }
