+++ mock/vault.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/evergreen-ci/cocoa/secret"
+	"github.com/evergreen-ci/cocoa"
 )
 
 // Vault provides a mock implementation of a secret.Vault. This makes it
@@ -15,7 +15,7 @@ type Vault struct{}
 // CreateSecret saves the input options and returns a mock secret ID. The mock
 // output can be customized. By default, it will create a cached mock secret
 // based on the input.
-func (m *Vault) CreateSecret(ctx context.Context, s secret.NamedSecret) (id string, err error) {
+func (m *Vault) CreateSecret(ctx context.Context, s cocoa.NamedSecret) (id string, err error) {
 	return "", errors.New("TODO: implement")
 }
