+++ secret/secrets_manager_vault.go
@@ -3,6 +3,7 @@ package secret
 import (
 	"context"
 
+	"github.com/evergreen-ci/cocoa"
 	"github.com/pkg/errors"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -12,18 +13,18 @@ import (
 // BasicSecretsManager provides a Vault implementation backed by Amazon Secrets
 // Manager.
 type BasicSecretsManager struct {
-	client SecretsManagerClient
+	client cocoa.SecretsManagerClient
 }
 
 // NewBasicSecretsManager creates a Vault backed by Secrets Manager.
-func NewBasicSecretsManager(c SecretsManagerClient) *BasicSecretsManager {
+func NewBasicSecretsManager(c cocoa.SecretsManagerClient) *BasicSecretsManager {
 	return &BasicSecretsManager{
 		client: c,
 	}
 }
 
 // CreateSecret creates a new secret.
-func (m *BasicSecretsManager) CreateSecret(ctx context.Context, s NamedSecret) (id string, err error) {
+func (m *BasicSecretsManager) CreateSecret(ctx context.Context, s cocoa.NamedSecret) (id string, err error) {
 	if err := s.Validate(); err != nil {
 		return "", errors.Wrap(err, "invalid secret")
 	}
